Added some styling, Poisson Matrix, Almost done.
Things to wrap up in the morning:

(1) how to get the Poisson JSON file loaded and then d3 to run it at
the beginning. Currently using jquery to open the json, but then I have
to figure out how to delay running the code until we get query to parse
it. It might be easier to wire the code to read the csv at output with
what we have?

(2) Make the SVG sit on top of the header bar so that the graph may
have some nodes on top of it (I think it would look cool).

(3) Better title?

(4) Another structured matrix would be cool.

(4)